move avatar image to core my pages
