+++ libs/portals/my-pages/core/src/index.ts
@@ -3,6 +3,7 @@ export * from './components/ActionCard/ActionCardLoader'
 export * from './components/ActionMenu/ActionMenu'
 export * from './components/ActionSidebar/ActionSidebar'
 export * from './components/AlertMessage/ModuleAlertMessageSection'
+export * from './components/AvatarImage/AvatarImage'
 export * from './components/AudioPlayer/AudioPlayer'
 export * from './components/CardLoader/CardLoader'
 export * from './components/Charts'
