+++ libs/portals/my-pages/documents/src/components/DocumentLine/DocumentLine.tsx
@@ -9,7 +9,11 @@ import {
 } from '@island.is/api/schema'
 import { Box, Icon, LoadingDots, Text, toast } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
-import { ConfirmationModal, m } from '@island.is/portals/my-pages/core'
+import {
+  AvatarImage,
+  ConfirmationModal,
+  m,
+} from '@island.is/portals/my-pages/core'
 import { dateFormat } from '@island.is/shared/constants'
 import {
   matchPath,
@@ -29,7 +33,6 @@ import {
 import { messages } from '../../utils/messages'
 import { FavAndStash } from '../FavAndStash/FavAndStash'
 import UrgentTag from '../UrgentTag/UrgentTag'
-import AvatarImage from './AvatarImage'
 import * as styles from './DocumentLine.css'
 import { Reply } from '../../lib/types'
 
@@ -309,6 +312,7 @@ export const DocumentLine: FC<Props> = ({
         <div ref={avatarRef} className={styles.avatar}>
           <AvatarImage
             img={img}
+            ariaLabel={formatMessage(messages.markAsBulkSelection)}
             onClick={(e) => {
               e.stopPropagation()
               if (documentLine.id && setSelectLine && !isUrgent) {
