+++ libs/portals/my-pages/documents/src/index.ts
@@ -2,5 +2,4 @@ export * from './module'
 export * from './lib/navigation'
 export * from './lib/paths'
 export * from './components/DocumentLine/DocumentLine'
-export * from './components/DocumentLine/AvatarImage'
 export * from './hooks/useDocumentList'
